+++ src/main.rs
@@ -8,6 +8,7 @@ use crate::portfolio::PortfolioArgs;
 use polars::prelude::*;
 use rust_decimal_macros::dec;
 use crate::markets::utils::save_candles;
+use crate::utils::print_candle_statistics;
 
 mod backtesting;
 mod indicators;
@@ -64,25 +65,7 @@ fn main() {
     let performance = runner.run(&candles, &market_data).unwrap();
     let elapsed = start_time.elapsed();
 
-    // print candle statistics
-    let candle_start = candles.column("time")
-        .unwrap()
-        .datetime()
-        .unwrap()
-        .head(Some(1))
-        .get(0)
-        .unwrap();
-    let candle_start = DateTime::from_timestamp_millis(candle_start).unwrap().naive_utc();
-    let candle_end = candles.column("time")
-        .unwrap()
-        .datetime()
-        .unwrap()
-        .tail(Some(1))
-        .get(0)
-        .unwrap();
-    let candle_end = DateTime::from_timestamp_millis(candle_end).unwrap().naive_utc();
-
-    info!("Candles range: {:?} - {:?}", candle_start, candle_end);
+    print_candle_statistics(&candles);
 
     let candle_len = candles.height();
     println!("Finished processing {:?} rows in {:?}", candle_len, elapsed);
@@ -95,9 +78,10 @@ fn main() {
 
     info!("Saving data to CSV");
 
+    let save_path = "data";
+
     let candle_path = format!("data/{}.csv", candle_table);
     let market_data_path = format!("data/{}.csv", market_data_table);
-    let bbands_path = "data/bbands_graph.csv";
 
     // create mutable versions of the DataFrames
     let mut candles = candles.clone();
@@ -108,6 +92,5 @@ fn main() {
     save_candles(&mut market_data, &market_data_path).unwrap();
 
     // save indicator graph as CSV
-    let mut bbands = runner.get_strategy_as_mut().indicators[0].as_mut();
-    let bbands_graph = bbands.save_graph_as_csv(bbands_path).unwrap();
+    runner.save_indicator_data(save_path).unwrap()
 }