+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add `data` dir to `.gitignore`
 - Create some utils to clean and better organize `main.rs` and `backtesting.rs`
 - Read backtesting config from a TOML file
+- Add documentation to `PositionManagerConfig` fields
 
 ---